Use cca3 for routing
